+++ src/frequenz/client/electricity_trading/cli/__main__.py
@@ -11,6 +11,12 @@ import click
 
 from frequenz.client.electricity_trading.cli.day_ahead import list_day_ahead_prices
 from frequenz.client.electricity_trading.cli.etrading import (
+    cancel_order as run_cancel_order,
+)
+from frequenz.client.electricity_trading.cli.etrading import (
+    create_order as run_create_order,
+)
+from frequenz.client.electricity_trading.cli.etrading import (
     list_orders as run_list_orders,
 )
 from frequenz.client.electricity_trading.cli.etrading import (
@@ -52,6 +58,70 @@ def list_trades(url: str, key: str, *, start: datetime) -> None:
 def list_orders(url: str, key: str, *, start: datetime, gid: int) -> None:
     """List orders."""
     asyncio.run(run_list_orders(url=url, key=key, delivery_start=start, gid=gid))
+
+
+@cli.command()
+@click.option("--url", required=True, type=str)
+@click.option("--key", required=True, type=str)
+@click.option("--start", required=True, type=iso)
+@click.option("--gid", required=True, type=int)
+@click.option("--quantity", required=True, type=str)
+@click.option("--price", required=True, type=str)
+@click.option("--area", required=True, type=str)
+@click.option("--currency", default="EUR", type=str)
+@click.option("--duration", default=900, type=int)
+def create_order(
+    # pylint: disable=too-many-arguments
+    url: str,
+    key: str,
+    *,
+    start: datetime,
+    gid: int,
+    quantity: str,
+    price: str,
+    area: str,
+    currency: str,
+    duration: int,
+) -> None:
+    """Create an order.
+
+    This is only allowed in test instances.
+    """
+    if "test" not in url:
+        raise ValueError("Creating orders is only allowed in test instances.")
+
+    asyncio.run(
+        run_create_order(
+            url=url,
+            key=key,
+            delivery_start=start,
+            gid=gid,
+            quantity_mw=quantity,
+            price=price,
+            delivery_area=area,
+            currency=currency,
+            duration=timedelta(seconds=duration),
+        )
+    )
+
+
+@cli.command()
+@click.option("--url", required=True, type=str)
+@click.option("--key", required=True, type=str)
+@click.option("--gid", required=True, type=int)
+@click.option("--order", required=True, type=int)
+def cancel_order(url: str, key: str, gid: int, order: int) -> None:
+    """Cancel an order."""
+    asyncio.run(run_cancel_order(url=url, key=key, gridpool_id=gid, order_id=order))
+
+
+@cli.command()
+@click.option("--url", required=True, type=str)
+@click.option("--key", required=True, type=str)
+@click.option("--gid", required=True, type=int)
+def cancel_all_orders(url: str, key: str, gid: int) -> None:
+    """Cancel all orders for a gridpool ID."""
+    asyncio.run(run_cancel_order(url=url, key=key, gridpool_id=gid, order_id=None))
 
 
 @cli.command()