+++ RELEASE_NOTES.md
@@ -14,6 +14,9 @@
   * `list-day-ahead`: Listing day-ahead prices from the entsoe API.
   * `list-trades`: Listing and streaming public trades for specified delivery periods. If no delivery start is given, starts streaming all new public trades.
   * `list-orders`: Listing and streaming orders for specified delivery periods and gridpool IDs. If no delivery start is given, starts streaming all new orders for this gridpool ID.
+  * `create-order`: Creating a limit order for a given price (in EUR/MWh) and quantity (in MW, sign determines market side).
+  * `cancel-order`: Cancel individual orders for a gridpool.
+  * `cancel-all-orders`: Cancels all orders of a gridpool.
 
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->